refactor: remove LunarPhaseDetailsRepo interface
