+++ app/src/main/java/dev/mslalith/focuslauncher/data/repository/LunarPhaseDetailsRepo.kt
@@ -1,4 +1,4 @@
-package dev.mslalith.focuslauncher.data.repository.interfaces
+package dev.mslalith.focuslauncher.data.repository
 
 import androidx.annotation.VisibleForTesting
 import dev.mslalith.focuslauncher.data.models.LunarPhase
@@ -7,21 +7,47 @@ import dev.mslalith.focuslauncher.data.models.LunarPhaseDirection
 import dev.mslalith.focuslauncher.data.models.Outcome
 import dev.mslalith.focuslauncher.data.models.UpcomingLunarPhase
 import dev.mslalith.focuslauncher.data.models.toLunarPhase
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.onEach
 import kotlinx.datetime.Instant
 import kotlinx.datetime.LocalDateTime
 import kotlinx.datetime.toJavaInstant
 import org.shredzone.commons.suncalc.MoonIllumination
 import org.shredzone.commons.suncalc.MoonPhase
+import javax.inject.Inject
+import kotlin.random.Random
 
-abstract class LunarPhaseDetailsRepo {
-    abstract val currentTimeStateFlow: StateFlow<Outcome<String>>
-    abstract val lunarPhaseDetailsStateFlow: StateFlow<Outcome<LunarPhaseDetails>>
-    abstract val upcomingLunarPhaseStateFlow: StateFlow<Outcome<UpcomingLunarPhase>>
+class LunarPhaseDetailsRepo @Inject constructor(
+    clockRepo: ClockRepo
+) {
+    private val _lunarPhaseDetailsStateFlow = MutableStateFlow<Outcome<LunarPhaseDetails>>(INITIAL_LUNAR_PHASE_DETAILS_OUTCOME)
+    val lunarPhaseDetailsStateFlow: StateFlow<Outcome<LunarPhaseDetails>>
+        get() = _lunarPhaseDetailsStateFlow
 
-    abstract fun refreshTime()
+    private val _upcomingLunarPhaseStateFlow = MutableStateFlow<Outcome<UpcomingLunarPhase>>(INITIAL_UPCOMING_LUNAR_PHASE_OUTCOME)
+    val upcomingLunarPhaseStateFlow: StateFlow<Outcome<UpcomingLunarPhase>>
+        get() = _upcomingLunarPhaseStateFlow
 
-    @VisibleForTesting(otherwise = VisibleForTesting.PROTECTED)
+    init {
+        clockRepo.currentInstantStateFlow.onEach { refreshLunarPhaseDetails(it) }
+    }
+
+    private suspend fun refreshLunarPhaseDetails(instant: Instant) {
+        val delayInMillis = Random.nextInt(from = 8, until = 17) * 100L
+        delay(delayInMillis)
+
+        val lunarPhaseDetails = findLunarPhaseDetails(instant)
+        updateStateFlowsWith(lunarPhaseDetails)
+    }
+
+    private fun updateStateFlowsWith(lunarPhaseDetails: LunarPhaseDetails) {
+        _lunarPhaseDetailsStateFlow.value = Outcome.Success(lunarPhaseDetails)
+        _upcomingLunarPhaseStateFlow.value = Outcome.Success(findUpcomingMoonPhaseFor(lunarPhaseDetails.direction))
+    }
+
+    @VisibleForTesting
     fun findLunarPhaseDetails(instant: Instant): LunarPhaseDetails =
         MoonIllumination.compute().on(instant.toJavaInstant()).execute().run {
             LunarPhaseDetails(
@@ -31,7 +57,7 @@ abstract class LunarPhaseDetailsRepo {
             )
         }
 
-    @VisibleForTesting(otherwise = VisibleForTesting.PROTECTED)
+    @VisibleForTesting
     fun findUpcomingMoonPhaseFor(lunarPhaseDirection: LunarPhaseDirection) =
         when (lunarPhaseDirection) {
             LunarPhaseDirection.NEW_TO_FULL -> findUpcomingLunarPhaseOf(LunarPhase.FULL_MOON)
@@ -52,5 +78,10 @@ abstract class LunarPhaseDetailsRepo {
             isMicroMoon = moonPhase.isMicroMoon,
             isSuperMoon = moonPhase.isSuperMoon,
         )
+    }
+
+    companion object {
+        val INITIAL_LUNAR_PHASE_DETAILS_OUTCOME = Outcome.Error("Has no Lunar Phase details")
+        val INITIAL_UPCOMING_LUNAR_PHASE_OUTCOME = Outcome.Error("Has no Upcoming Lunar Phase")
     }
 }
